requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.0
deprecated==1.2.15
filelock==3.13.1
flatbuffers==24.3.25
fsspec==2024.2.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
inquirerpy==0.3.4
jinja2==3.1.3
markupsafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.2.0
onnx==1.17.0
onnxconverter-common==1.14.0
onnxruntime==1.20.1
onnxruntime-tools==1.7.0
packaging==24.2
pfzy==0.3.4
prompt-toolkit==3.0.48
protobuf==3.20.2
psutil==6.1.0
py-cpuinfo==9.0.0
py3nvml==0.2.7
pycparser==2.22
pygithub==2.5.0
pyjwt==2.10.1
pynacl==1.5.0
pyreadline3==3.5.4
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing-extensions==4.9.0
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.17.0
xmltodict==0.14.2
